Added all Services to the frontend. But they need to programmed again, because the backednd has problems with the Jpa models.
Extra: All endpoits are opened for testing. The security gets testet at the end of development.

The Frontend got extended by angular materials for the pagiantion.

The rooting does som problem with the shop page -> Info: could be possible that the frontend domain routing collides with the backend domian routing.